Update library/Zend/Http/Header/SetCookie.php
Cookies were not handled correctly. If you had a cookie for .sub.domain.de from the last response and then fired a new request to www.sub.domain.de, the old cookie was not added to the request (because .sub.domain.de is a substring of www.sub.domain.de, it should be added, the method should return true which was not the case).