HTML5 compliant form attributes
According to W3C specification, form ‘action’ attribute can’t be empty in HTML5 — but now, is optional. So, if the form doctype is ‘HTML5’ or ‘XHTML5’, the form action isn’t set as an empty string. Also, if either doctypes are detected, form ‘method’ attribute is not set to ‘get’ by default, as some applications may use only ‘formmethod’ attribute on submit buttons (according to W3C, “missing value default for the method attribute is also the GET state” — and this is valid since HTML 2.0).

Reference:
- http://www.w3.org/TR/html5/forms.html#attr-fs-action
- http://www.w3.org/TR/html5/forms.html#attr-fs-method