Adjust ContentType regex to allow quoted strings and equals
This change modifies the regular expression in ContentType::fromString()
to allow parameters to contain the equals (=) character, and be
optionally enclosed in double quotes. For example the following header
is perfectly valid, but does not get currently parsed by ContentType:

  Content-type: text/plain; charset="us-ascii"

Whilst allowing the parameter to be enclosed in double quotes, it should
not be included in the value itself, so these are outside the capture
group in the regular expression.

This change should resolve issue #7076.